backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
Werkzeug==2.3.7
python-dotenv==1.0.0
PyJWT==2.8.0
psycopg2-binary==2.9.7
python-socketio==5.8.0
eventlet==0.33.3
pytest==7.4.3
pytest-flask==1.2.0
coverage==7.3.2